backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
sqlalchemy>=2.0
pandas>=2.0
duckdb>=1.0
openpyxl>=3.1
python-multipart>=0.0.6
openai>=1.0
python-dotenv>=1.0
pydantic>=2.0
psycopg2-binary
pymysql
aiofiles>=23.0
alembic>=1.13.0
anthropic>=0.18.0
litellm>=1.0
cryptography
tabulate
